Simple command framework with first command
Base project files with a basic framework to quickly add new
subcommands with options to allow building some automation with the
various services to assist with the organization of Devopsdays Galway
conference.

Includes a basic command to retrieve cards in progress from the trello
board to generate an agenda for sync up meetings.