Command to create RSVP tickets using generic email
Adds a simple subcommand to create RSVP tickets when using a generic
email address using the desired settings, allowing for control of which
ticket type would be included and the name of the list.

Also support outputting the unique urls in the case the are already
existing tickets in the RSVP list to make it easy to gather all links in
subsequent lines.